requirements.txt
streamlit
pandas
numpy
plotly
lightgbm
scikit-learn
chronos-forecasting==2.2.0
